Fix progress status

+++ generate_test_file.py
@@ -129,7 +129,7 @@ with open('clue.json') as f, open('clue_anomaly.json', 'w+') as out, open('label
     for line in f:
         cnt += 1
         buf_cnt += 1
-        if cnt % (total_lines / 20) == 0:
+        if int(cnt % (total_lines / 20)) == 0: 
             print(str(int(cnt*100/total_lines)) + '%', end=' ', flush=True)
         j = json.loads(line)
         uid = j['uid']
